Fixing parameter types to accomodate resizable canvases

+++ src/modules/graphics.c
@@ -2,8 +2,8 @@ internal void
 CANVAS_print(WrenVM* vm) {
   ENGINE* engine = (ENGINE*)wrenGetUserData(vm);
   char* text = (char*)wrenGetSlotString(vm, 1);
-  int16_t x = round(wrenGetSlotDouble(vm, 2));
-  int16_t y = round(wrenGetSlotDouble(vm, 3));
+  int64_t x = round(wrenGetSlotDouble(vm, 2));
+  int64_t y = round(wrenGetSlotDouble(vm, 3));
   uint32_t c = round(wrenGetSlotDouble(vm, 4));
 
   ENGINE_print(engine, text, x, y, c);
@@ -13,8 +13,8 @@ internal void
 CANVAS_pset(WrenVM* vm)
 {
   ENGINE* engine = (ENGINE*)wrenGetUserData(vm);
-  int16_t x = round(wrenGetSlotDouble(vm, 1));
-  int16_t y = round(wrenGetSlotDouble(vm, 2));
+  int64_t x = round(wrenGetSlotDouble(vm, 1));
+  int64_t y = round(wrenGetSlotDouble(vm, 2));
   uint32_t c = round(wrenGetSlotDouble(vm, 3));
   ENGINE_pset(engine, x,y,c);
 }
@@ -23,9 +23,9 @@ internal void
 CANVAS_circle_filled(WrenVM* vm)
 {
   ENGINE* engine = (ENGINE*)wrenGetUserData(vm);
-  int16_t x = round(wrenGetSlotDouble(vm, 1));
-  int16_t y = round(wrenGetSlotDouble(vm, 2));
-  int16_t r = round(wrenGetSlotDouble(vm, 3));
+  int64_t x = round(wrenGetSlotDouble(vm, 1));
+  int64_t y = round(wrenGetSlotDouble(vm, 2));
+  int64_t r = round(wrenGetSlotDouble(vm, 3));
   uint32_t c = round(wrenGetSlotDouble(vm, 4));
   ENGINE_circle_filled(engine, x, y, r, c);
 }
@@ -34,9 +34,9 @@ internal void
 CANVAS_circle(WrenVM* vm)
 {
   ENGINE* engine = (ENGINE*)wrenGetUserData(vm);
-  int16_t x = round(wrenGetSlotDouble(vm, 1));
-  int16_t y = round(wrenGetSlotDouble(vm, 2));
-  int16_t r = round(wrenGetSlotDouble(vm, 3));
+  int64_t x = round(wrenGetSlotDouble(vm, 1));
+  int64_t y = round(wrenGetSlotDouble(vm, 2));
+  int64_t r = round(wrenGetSlotDouble(vm, 3));
   uint32_t c = round(wrenGetSlotDouble(vm, 4));
   ENGINE_circle(engine, x, y, r, c);
 }
@@ -44,10 +44,10 @@ internal void
 CANVAS_line(WrenVM* vm)
 {
   ENGINE* engine = (ENGINE*)wrenGetUserData(vm);
-  int16_t x1 = round(wrenGetSlotDouble(vm, 1));
-  int16_t y1 = round(wrenGetSlotDouble(vm, 2));
-  int16_t x2 = round(wrenGetSlotDouble(vm, 3));
-  int16_t y2 = round(wrenGetSlotDouble(vm, 4));
+  int64_t x1 = round(wrenGetSlotDouble(vm, 1));
+  int64_t y1 = round(wrenGetSlotDouble(vm, 2));
+  int64_t x2 = round(wrenGetSlotDouble(vm, 3));
+  int64_t y2 = round(wrenGetSlotDouble(vm, 4));
   uint32_t c = round(wrenGetSlotDouble(vm, 5));
   ENGINE_line(engine, x1, y1, x2, y2, c);
 }
@@ -56,10 +56,10 @@ internal void
 CANVAS_ellipse(WrenVM* vm)
 {
   ENGINE* engine = (ENGINE*)wrenGetUserData(vm);
-  int16_t x1 = round(wrenGetSlotDouble(vm, 1));
-  int16_t y1 = round(wrenGetSlotDouble(vm, 2));
-  int16_t x2 = round(wrenGetSlotDouble(vm, 3));
-  int16_t y2 = round(wrenGetSlotDouble(vm, 4));
+  int64_t x1 = round(wrenGetSlotDouble(vm, 1));
+  int64_t y1 = round(wrenGetSlotDouble(vm, 2));
+  int64_t x2 = round(wrenGetSlotDouble(vm, 3));
+  int64_t y2 = round(wrenGetSlotDouble(vm, 4));
   uint32_t c = round(wrenGetSlotDouble(vm, 5));
   ENGINE_ellipse(engine, x1, y1, x2, y2, c);
 }
@@ -68,10 +68,10 @@ internal void
 CANVAS_ellipsefill(WrenVM* vm)
 {
   ENGINE* engine = (ENGINE*)wrenGetUserData(vm);
-  int16_t x1 = round(wrenGetSlotDouble(vm, 1));
-  int16_t y1 = round(wrenGetSlotDouble(vm, 2));
-  int16_t x2 = round(wrenGetSlotDouble(vm, 3));
-  int16_t y2 = round(wrenGetSlotDouble(vm, 4));
+  int64_t x1 = round(wrenGetSlotDouble(vm, 1));
+  int64_t y1 = round(wrenGetSlotDouble(vm, 2));
+  int64_t x2 = round(wrenGetSlotDouble(vm, 3));
+  int64_t y2 = round(wrenGetSlotDouble(vm, 4));
   uint32_t c = round(wrenGetSlotDouble(vm, 5));
   ENGINE_ellipsefill(engine, x1, y1, x2, y2, c);
 }
@@ -80,10 +80,10 @@ internal void
 CANVAS_rect(WrenVM* vm)
 {
   ENGINE* engine = (ENGINE*)wrenGetUserData(vm);
-  int16_t x = round(wrenGetSlotDouble(vm, 1));
-  int16_t y = round(wrenGetSlotDouble(vm, 2));
-  int16_t w = round(wrenGetSlotDouble(vm, 3));
-  int16_t h = round(wrenGetSlotDouble(vm, 4));
+  int64_t x = round(wrenGetSlotDouble(vm, 1));
+  int64_t y = round(wrenGetSlotDouble(vm, 2));
+  int64_t w = round(wrenGetSlotDouble(vm, 3));
+  int64_t h = round(wrenGetSlotDouble(vm, 4));
   uint32_t c = round(wrenGetSlotDouble(vm, 5));
   ENGINE_rect(engine, x, y, w, h, c);
 }
@@ -92,10 +92,10 @@ internal void
 CANVAS_rectfill(WrenVM* vm)
 {
   ENGINE* engine = (ENGINE*)wrenGetUserData(vm);
-  int16_t x = round(wrenGetSlotDouble(vm, 1));
-  int16_t y = round(wrenGetSlotDouble(vm, 2));
-  int16_t w = round(wrenGetSlotDouble(vm, 3));
-  int16_t h = round(wrenGetSlotDouble(vm, 4));
+  int64_t x = round(wrenGetSlotDouble(vm, 1));
+  int64_t y = round(wrenGetSlotDouble(vm, 2));
+  int64_t w = round(wrenGetSlotDouble(vm, 3));
+  int64_t h = round(wrenGetSlotDouble(vm, 4));
   uint32_t c = round(wrenGetSlotDouble(vm, 5));
   ENGINE_rectfill(engine, x, y, w, h, c);
 }
